Crate moon

Re-exports

pub use actix_cors;

pub use actix_files;

pub use actix_http;

pub use actix_web;

pub use chashmap;

pub use futures;

pub use mime;

pub use mime_guess;

pub use moonlight;

pub use once_cell;

pub use parking_lot;

pub use rustls;

pub use rustls_pemfile;

pub use serde;

pub use tokio;

pub use tokio_stream;

pub use uuid;

Modules

chrono

Chrono: Date and Time for Rust

config

error_handler

rusty_ulid

ULID - Universally Unique Lexicographically Sortable Identifier

serde

Serde

serde_json

Serde JSON

sessions

Macros

clone

Macro for cloning values to close. Alternative short record.

trait_set

Creates an alias for set of traits.

Structs

ActorId

AuthToken

CorId

Date

ISO 8601 calendar date with time zone.

DateTime

ISO 8601 combined date and time with time zone.

DownMsgTransporterForDe

DownMsgTransporterForSer

Duration

ISO 8601 time duration with nanosecond precision. This also allows for the negative duration; see individual methods for details.

EntityId

FixedOffset

The time zone with fixed offset, from UTC-23:59:59 to UTC+23:59:59.

Frontend

Lazy

A value which is initialized on the first access.

Local

The local timescale. This is implemented via the standard time crate.

MessageSSE

NaiveDate

ISO 8601 calendar date without timezone. Allows for every proleptic Gregorian date from Jan 1, 262145 BCE to Dec 31, 262143 CE. Also supports the conversion from ISO 8601 ordinal and week date.

NaiveDateTime

ISO 8601 combined date and time without timezone.

NaiveTime

ISO 8601 time without timezone. Allows for the nanosecond precision and optional leap second representation.

Redirect

SessionActor

SessionId

Ulid

The ULID data type.

UpMsgRequest

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

Wrapper

Enums

DecodingError

Error that can occur while decoding a crockford Base32 string.

Lang

Month

The month of the year.

SecondsFormat

Specific formatting options for seconds. This may be extended in the future, so exhaustive matching in external code is not recommended.

Weekday

The day of week.

Traits

ActorInstance

Also

Represents a type that you can apply arbitrary functions to. Useful for when a method doesn’t return the receiver but you want to apply several of them to the object.

Apply

Represents a type which can have functions applied to it (implemented by default for all types).

Datelike

The common set of methods for date component.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned

A data structure that can be deserialized without borrowing any data from the deserializer.

Deserializer

A data format that can deserialize any data structure supported by Serde.

FromEnvVars

FrontBuilder

FrontBuilderOutput

Index

Offset

The offset from the local time to UTC.

PVar

Serialize

A data structure that can be serialized into any data format supported by Serde.

Serializer

A data format that can serialize any data structure supported by Serde.

SubsecRound

Extension trait for subsecond rounding or truncation to a maximum number of digits. Rounding can be used to decrease the error variance when serializing/persisting to lower precision. Truncation is the default behavior in Chrono display formatting. Either can be used to guarantee equality (e.g. for testing) when round-tripping through a lower precision format.

TimeZone

The time zone.

Timelike

The common set of methods for time component.

UpHandler

UpHandlerOutput

Functions

not

start

start_with_app

Attribute Macros

async_trait

main

Marks async main function as the Moon system entry-point.

test

Marks async test functions to use the Moon system entry-point.

Derive Macros

Deserialize

Serialize